refactor: simplify recovery workflows and harden saving
