+++ src/Beutl.Editor/AutoSaveService.cs
@@ -33,8 +33,20 @@ public sealed class AutoSaveService : IDisposable
     public void SaveObjects(IEnumerable<CoreObject> objectsToSave)
     {
         ThrowIfDisposed();
+        CoreObject[] objects = objectsToSave.ToArray();
+        try
+        {
+            CoreSerializer.PersistProjectMigrationMetadata(objects);
+        }
+        catch (Exception ex)
+        {
+            _logger.LogError(ex, "Could not persist project migration requirements; auto-save was aborted.");
+            _saveError.OnNext(ex);
+            return;
+        }
+
         // 各オブジェクトを保存
-        foreach (CoreObject obj in objectsToSave)
+        foreach (CoreObject obj in objects)
         {
             try
             {
