+++ src/Beutl/Services/AI/AiVideoResultDownload.cs
@@ -1,7 +1,53 @@
-﻿namespace Beutl.Services.AI;
+﻿using Beutl.Api.Services;
+using Beutl.Logging;
+using Microsoft.Extensions.Logging;
+
+namespace Beutl.Services.AI;
 
 internal static class AiVideoResultDownload
 {
+    private static readonly ILogger s_logger = Log.CreateLogger(typeof(AiVideoResultDownload));
+
+    public static async Task<string> DownloadAsync(
+        IAuthenticatedContentService content,
+        Uri contentUri,
+        AiContentMetadata? declaredMetadata,
+        CancellationToken cancellationToken)
+    {
+        (string stagingPath, FileStream destination) = AiTemporaryFileStore.Create(
+            "results", "ai-video", ".download");
+        string ownedPath = stagingPath;
+        try
+        {
+            AiContentDownload download;
+            await using (destination)
+            {
+                using Stream bounded = CreateBoundedStream(destination);
+                download = await content.CopyToAsync(contentUri, bounded, cancellationToken);
+            }
+
+            AiContentMetadata? metadata = AiContentMetadata.Combine(declaredMetadata, download.Metadata);
+            string extension = metadata?.GetFileExtension(".mp4", "video") ?? ".mp4";
+            string completedPath = Path.ChangeExtension(stagingPath, extension);
+            File.Move(stagingPath, completedPath);
+            ownedPath = completedPath;
+            AiTemporaryFileStore.EnsurePrivateFile(completedPath);
+            return completedPath;
+        }
+        catch
+        {
+            try
+            {
+                File.Delete(ownedPath);
+            }
+            catch (Exception ex) when (ex is IOException or UnauthorizedAccessException)
+            {
+                s_logger.LogWarning(ex, "Failed to remove incomplete AI video {Path}.", ownedPath);
+            }
+            throw;
+        }
+    }
+
     // The API accepts at most 60 seconds. This still permits an encoded bitrate above 68 Mbit/s
     // while bounding a malformed or non-terminating response well below the whole disk.
     internal const long MaximumBytes = 512L * 1024 * 1024;
