+++ src/Beutl/ViewModels/Dialogs/AiSubtitleDialogViewModel.Advanced.cs
@@ -3,7 +3,6 @@ using System.Collections.Specialized;
 using System.ComponentModel;
 using System.Globalization;
 using System.Reactive.Disposables;
-using System.Text;
 using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Controls.ApplicationLifetimes;
@@ -25,6 +24,9 @@ using Beutl.Services.AI;
 using FluentAvalonia.UI.Controls;
 using Microsoft.Extensions.Logging;
 using Reactive.Bindings;
+
+using static Beutl.Services.AI.CaptionTranslationBatcher;
+using static Beutl.Services.AI.SpeechWaveEncoder;
 
 namespace Beutl.ViewModels.Dialogs;
 
@@ -1459,7 +1461,7 @@ public sealed partial class AiSubtitleDialogViewModel
             targetLanguage,
             draftScopeRevision,
             limits,
-            CreateTranslationBatches(
+            CaptionTranslationBatcher.CreateBatches(
                 sourceDocument,
                 sourceLanguage,
                 targetLanguage,
@@ -1479,7 +1481,7 @@ public sealed partial class AiSubtitleDialogViewModel
         AiModelId? model = string.IsNullOrEmpty(resume.RequestKeyModel)
             ? null
             : new AiModelId(resume.RequestKeyModel);
-        List<TranslationBatch> batches = CreateTranslationBatches(
+        List<TranslationBatch> batches = CaptionTranslationBatcher.CreateBatches(
             sourceDocument,
             resume.SourceLanguage,
             resume.TargetLanguage,
@@ -2678,7 +2680,7 @@ public sealed partial class AiSubtitleDialogViewModel
                 codec.Format,
                 _captionSerializer,
                 ConfirmLossySrtExportAsync,
-                (bytes, cancellationToken) => WriteCaptionExportAsync(
+                (bytes, cancellationToken) => CaptionExportStorage.WriteAsync(
                     file,
                     bytes,
                     cancellationToken),
@@ -3417,7 +3419,7 @@ public sealed partial class AiSubtitleDialogViewModel
         {
             AiCaptionTranslationLimits limits =
                 TranslationModelPicker.CaptionTranslationLimits;
-            request = CreateTranslationBatches(
+            request = CaptionTranslationBatcher.CreateBatches(
                     document,
                     SelectedSourceLanguage.Value.Code ?? _lastCaptionLanguage,
                     targetLanguage,
@@ -3900,713 +3902,6 @@ public sealed partial class AiSubtitleDialogViewModel
             ? null
             : string.Format(Strings.AiSubtitle_DetectedLanguage, language);
 
-    private static List<TranslationBatch> CreateTranslationBatches(
-        CaptionDocument document,
-        string? sourceLanguage,
-        string targetLanguage,
-        AiModelId? model,
-        AiCaptionTranslationLimits limits)
-    {
-        var pieces = new List<TranslationPiece>();
-        for (int cueIndex = 0; cueIndex < document.Count; cueIndex++)
-        {
-            CaptionCue cue = document[cueIndex];
-            if (string.IsNullOrWhiteSpace(cue.Text))
-                continue;
-
-            int[] textElementBoundaries = CreateTranslationTextElementBoundaries(cue.Text);
-            int offset = 0;
-            int partIndex = 0;
-            while (offset < cue.Text.Length)
-            {
-                int maximumLength = Math.Min(
-                    limits.MaxCharacters,
-                    cue.Text.Length - offset);
-                int length = LargestTranslationPieceLength(
-                    cue,
-                    cueIndex,
-                    partIndex,
-                    offset,
-                    maximumLength,
-                    textElementBoundaries,
-                    sourceLanguage,
-                    targetLanguage,
-                    model,
-                    limits);
-                if (length <= 0)
-                    throw new SubtitleInputException(Strings.AiFileTooLarge);
-
-                pieces.Add(CreateTranslationPiece(
-                    cue,
-                    cueIndex,
-                    partIndex,
-                    cue.Text.Substring(offset, length),
-                    limits));
-                offset += length;
-                partIndex++;
-            }
-        }
-
-        var batches = new List<TranslationBatch>();
-        int start = 0;
-        while (start < pieces.Count)
-        {
-            int candidateCount = 0;
-            int characters = 0;
-            while (start + candidateCount < pieces.Count
-                   && candidateCount < limits.MaxSegments
-                   && characters + pieces[start + candidateCount].Text.Length
-                   <= limits.MaxCharacters)
-            {
-                characters += pieces[start + candidateCount].Text.Length;
-                candidateCount++;
-            }
-
-            int count = LargestTranslationBatchPrefix(
-                pieces,
-                start,
-                candidateCount,
-                sourceLanguage,
-                targetLanguage,
-                model,
-                limits);
-            if (count <= 0)
-                throw new SubtitleInputException(Strings.AiFileTooLarge);
-
-            batches.Add(new TranslationBatch(pieces.GetRange(start, count).ToArray()));
-            start += count;
-        }
-        return batches;
-    }
-
-    private static int LargestTranslationPieceLength(
-        CaptionCue cue,
-        int cueIndex,
-        int partIndex,
-        int offset,
-        int maximumLength,
-        int[] textElementBoundaries,
-        string? sourceLanguage,
-        string targetLanguage,
-        AiModelId? model,
-        AiCaptionTranslationLimits limits)
-    {
-        int low = 1;
-        int high = maximumLength;
-        int best = 0;
-        while (low <= high)
-        {
-            int midpoint = low + ((high - low) / 2);
-            int length = KeepTranslationBoundaryTogether(
-                cue.Text,
-                textElementBoundaries,
-                offset,
-                midpoint);
-            if (length <= 0)
-            {
-                low = midpoint + 1;
-                continue;
-            }
-
-            TranslationPiece piece = CreateTranslationPiece(
-                cue,
-                cueIndex,
-                partIndex,
-                cue.Text.Substring(offset, length),
-                limits);
-            if (TranslationBatchFits(
-                    [piece],
-                    sourceLanguage,
-                    targetLanguage,
-                    model,
-                    limits))
-            {
-                best = Math.Max(best, length);
-                low = midpoint + 1;
-            }
-            else
-            {
-                high = midpoint - 1;
-            }
-        }
-        return best;
-    }
-
-    private static int LargestTranslationBatchPrefix(
-        List<TranslationPiece> pieces,
-        int start,
-        int maximumCount,
-        string? sourceLanguage,
-        string targetLanguage,
-        AiModelId? model,
-        AiCaptionTranslationLimits limits)
-    {
-        int low = 1;
-        int high = maximumCount;
-        int best = 0;
-        while (low <= high)
-        {
-            int count = low + ((high - low) / 2);
-            if (TranslationBatchFits(
-                    pieces.GetRange(start, count),
-                    sourceLanguage,
-                    targetLanguage,
-                    model,
-                    limits))
-            {
-                best = count;
-                low = count + 1;
-            }
-            else
-            {
-                high = count - 1;
-            }
-        }
-        return best;
-    }
-
-    private static bool TranslationBatchFits(
-        IReadOnlyList<TranslationPiece> pieces,
-        string? sourceLanguage,
-        string targetLanguage,
-        AiModelId? model,
-        AiCaptionTranslationLimits limits)
-    {
-        try
-        {
-            _ = new AiCaptionTranslationRequest(
-                pieces.Select(piece => new AiCaptionTranslationSegment
-                {
-                    Id = piece.Id,
-                    Text = piece.Text,
-                    Context = new AiCaptionTranslationSegmentContext(
-                        piece.GroupId,
-                        piece.ContextPartIndex,
-                        piece.Start,
-                        piece.End),
-                }).ToArray(),
-                targetLanguage,
-                sourceLanguage,
-                model: model,
-                limits: limits);
-            return true;
-        }
-        catch (ArgumentException)
-        {
-            return false;
-        }
-    }
-
-    private static TranslationPiece CreateTranslationPiece(
-        CaptionCue cue,
-        int cueIndex,
-        int partIndex,
-        string text,
-        AiCaptionTranslationLimits limits)
-    {
-        int contextGroup = partIndex / limits.MaxSegments;
-        return new TranslationPiece(
-            cueIndex,
-            partIndex,
-            partIndex % limits.MaxSegments,
-            $"c{cueIndex}-p{partIndex}",
-            contextGroup == 0 ? $"c{cueIndex}" : $"c{cueIndex}-g{contextGroup}",
-            cue.Start,
-            cue.End,
-            text);
-    }
-
-    internal static int[] CreateTranslationTextElementBoundaries(string text)
-        => StringInfo.ParseCombiningCharacters(text);
-
-    internal static int KeepTranslationBoundaryTogether(
-        string text,
-        int[] textElementBoundaries,
-        int offset,
-        int length)
-    {
-        int end = offset + length;
-        if (end >= text.Length || length <= 0)
-            return length;
-
-        int boundaryIndex = Array.BinarySearch(textElementBoundaries, end);
-        if (boundaryIndex < 0)
-            boundaryIndex = ~boundaryIndex - 1;
-        if (boundaryIndex < 0 || textElementBoundaries[boundaryIndex] <= offset)
-            return 0;
-        int textElementBoundary = textElementBoundaries[boundaryIndex];
-        int offsetBoundaryIndex = Array.BinarySearch(textElementBoundaries, offset);
-        if (offsetBoundaryIndex < 0)
-            throw new ArgumentException("The offset must be a text-element boundary.", nameof(offset));
-
-        for (int candidateIndex = boundaryIndex;
-             candidateIndex > offsetBoundaryIndex;
-             candidateIndex--)
-        {
-            int boundary = textElementBoundaries[candidateIndex];
-            int previousTextElement = textElementBoundaries[candidateIndex - 1];
-            if (!char.IsWhiteSpace(text[previousTextElement])
-                || (boundary < text.Length && char.IsWhiteSpace(text[boundary])))
-            {
-                continue;
-            }
-
-            int separatorIndex = candidateIndex - 1;
-            while (separatorIndex > offsetBoundaryIndex
-                   && char.IsWhiteSpace(text[textElementBoundaries[separatorIndex - 1]]))
-            {
-                separatorIndex--;
-            }
-            int separatorStart = textElementBoundaries[separatorIndex];
-            if (separatorStart > offset)
-            {
-                // Keep the complete separator with the preceding word. A
-                // leading separator is part of the next word for this purpose.
-                return boundary - offset;
-            }
-        }
-
-        int trailingTextElement = textElementBoundaries[boundaryIndex - 1];
-        if (char.IsWhiteSpace(text[trailingTextElement])
-            && char.IsWhiteSpace(text[textElementBoundary]))
-        {
-            int separatorIndex = boundaryIndex - 1;
-            while (separatorIndex > offsetBoundaryIndex
-                   && char.IsWhiteSpace(text[textElementBoundaries[separatorIndex - 1]]))
-            {
-                separatorIndex--;
-            }
-            int separatorStart = textElementBoundaries[separatorIndex];
-            if (separatorStart > offset)
-                return separatorStart - offset;
-        }
-
-        // A single word longer than the provider limit still has to make
-        // progress, but it may only be split between complete text elements.
-        return textElementBoundary - offset;
-    }
-
-    internal static SpeechWaveChunkResult WriteSpeechWave(
-        MediaReader reader,
-        int startSample,
-        int requestedSamples,
-        Stream stream,
-        CancellationToken cancellationToken)
-    {
-        ArgumentNullException.ThrowIfNull(reader);
-        ArgumentNullException.ThrowIfNull(stream);
-        if (startSample < 0)
-            throw new ArgumentOutOfRangeException(nameof(startSample));
-        if (requestedSamples <= 0)
-            throw new ArgumentOutOfRangeException(nameof(requestedSamples));
-        cancellationToken.ThrowIfCancellationRequested();
-        if (!stream.CanWrite || !stream.CanSeek)
-        {
-            throw new ArgumentException(
-                "The destination stream must be writable and seekable.",
-                nameof(stream));
-        }
-
-        int sourceRate = reader.AudioInfo.SampleRate;
-        if (sourceRate <= 0)
-            throw new SubtitleInputException(Strings.AiSubtitle_NoAudioInRange);
-
-        const int outputRate = 16_000;
-        const int waveHeaderLength = 44;
-        using var writer = new BinaryWriter(stream, Encoding.ASCII, leaveOpen: true);
-        writer.Write(Encoding.ASCII.GetBytes("RIFF"));
-        writer.Write(0);
-        writer.Write(Encoding.ASCII.GetBytes("WAVEfmt "));
-        writer.Write(16);
-        writer.Write((short)1);
-        writer.Write((short)1);
-        writer.Write(outputRate);
-        writer.Write(outputRate * sizeof(short));
-        writer.Write((short)sizeof(short));
-        writer.Write((short)16);
-        writer.Write(Encoding.ASCII.GetBytes("data"));
-        writer.Write(0);
-
-        int sourceSamplesWritten = 0;
-        int outputSamplesWritten = 0;
-        int decodeBlockSize = Math.Max(1, Math.Min(sourceRate, requestedSamples));
-        double nextOutputSourcePosition = 0;
-        while (sourceSamplesWritten < requestedSamples)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-            int length = Math.Min(decodeBlockSize, requestedSamples - sourceSamplesWritten);
-            if (!reader.ReadAudio(
-                    checked(startSample + sourceSamplesWritten),
-                    length,
-                    out Ref<IPcm>? audioRef))
-            {
-                throw new SubtitleInputException(Strings.AiSubtitle_NoAudioInRange);
-            }
-
-            using (audioRef)
-            {
-                if (audioRef.Value is not Pcm<Stereo32BitFloat> pcm)
-                    throw new SubtitleInputException(Strings.AiSubtitle_NoAudioInRange);
-
-                int decodedSamples = Math.Min(pcm.NumSamples, length);
-                if (decodedSamples == 0)
-                    break;
-
-                Span<Stereo32BitFloat> source = pcm.DataSpan[..decodedSamples];
-                double blockEnd = sourceSamplesWritten + decodedSamples;
-                while (nextOutputSourcePosition < blockEnd)
-                {
-                    int sourceIndex = Math.Clamp(
-                        (int)Math.Floor(nextOutputSourcePosition - sourceSamplesWritten),
-                        0,
-                        decodedSamples - 1);
-                    Stereo32BitFloat sample = source[sourceIndex];
-                    double mono = (sample.Left + sample.Right) / 2d;
-                    mono = double.IsFinite(mono) ? Math.Clamp(mono, -1, 1) : 0;
-                    writer.Write((short)Math.Round(mono * short.MaxValue));
-                    outputSamplesWritten++;
-                    nextOutputSourcePosition = outputSamplesWritten * (double)sourceRate / outputRate;
-                }
-                sourceSamplesWritten += decodedSamples;
-                if (decodedSamples < length)
-                    break;
-            }
-        }
-
-        if (sourceSamplesWritten == 0 || outputSamplesWritten == 0)
-            throw new SubtitleInputException(Strings.AiSubtitle_NoAudioInRange);
-
-        int dataLength = checked(outputSamplesWritten * sizeof(short));
-        stream.Position = 4;
-        writer.Write(checked(waveHeaderLength - 8 + dataLength));
-        stream.Position = 40;
-        writer.Write(dataLength);
-        stream.Position = waveHeaderLength + dataLength;
-        return new SpeechWaveChunkResult(
-            sourceSamplesWritten,
-            outputSamplesWritten,
-            TimeSpan.FromSeconds(outputSamplesWritten / (double)outputRate));
-    }
-
-    /// <summary>
-    /// Writes one 16-bit mono PCM wave from as many composed slices as it is
-    /// handed. The slices are separate calls because a whole upload chunk cannot
-    /// be composed at once (see <see cref="SceneMixComposeSlice"/>), and the
-    /// resampling position carries across them so a slice boundary neither drops
-    /// a sample nor shifts the ones after it.
-    /// </summary>
-    internal sealed class SpeechWaveWriter(Stream stream)
-    {
-        private const int WaveHeaderLength = 44;
-        private const int MaximumOutputRate = 16_000;
-        private readonly Stream _stream = stream;
-        private int _sourceRate;
-        private int _outputRate;
-        private long _sourceFramesWritten;
-        private double _nextOutputSourcePosition;
-        private int _outputSampleCount;
-
-        public TimeSpan UploadedDuration => _outputRate == 0
-            ? TimeSpan.Zero
-            : TimeSpan.FromSeconds(_outputSampleCount / (double)_outputRate);
-
-        public void Append(AudioFrameSnapshot snapshot, CancellationToken cancellationToken)
-        {
-            ArgumentNullException.ThrowIfNull(snapshot);
-            cancellationToken.ThrowIfCancellationRequested();
-            if (snapshot.SampleRate <= 0 || snapshot.ChannelCount <= 0 || snapshot.SampleCount <= 0)
-                throw new SubtitleInputException(Strings.AiSubtitle_NoAudioInRange);
-
-            if (_sourceRate == 0)
-            {
-                _sourceRate = snapshot.SampleRate;
-                _outputRate = Math.Min(_sourceRate, MaximumOutputRate);
-                WriteHeader();
-            }
-            else if (snapshot.SampleRate != _sourceRate)
-            {
-                // Resampling assumes one rate for the whole wave; a slice at another
-                // rate would land its samples at the wrong time.
-                throw new SubtitleInputException(Strings.AiSubtitle_NoAudioInRange);
-            }
-
-            using var writer = new BinaryWriter(_stream, Encoding.ASCII, leaveOpen: true);
-            long sliceStart = _sourceFramesWritten;
-            long sliceEnd = sliceStart + snapshot.SampleCount;
-            while (_nextOutputSourcePosition < sliceEnd)
-            {
-                if ((_outputSampleCount & 4095) == 0)
-                    cancellationToken.ThrowIfCancellationRequested();
-
-                int frame = (int)Math.Clamp(
-                    (long)Math.Floor(_nextOutputSourcePosition) - sliceStart,
-                    0,
-                    snapshot.SampleCount - 1);
-                double mono = 0;
-                for (int channel = 0; channel < snapshot.ChannelCount; channel++)
-                {
-                    float sample = snapshot.Interleaved[(frame * snapshot.ChannelCount) + channel];
-                    mono += float.IsFinite(sample) ? sample : 0;
-                }
-
-                mono = Math.Clamp(mono / snapshot.ChannelCount, -1, 1);
-                writer.Write((short)Math.Round(mono * short.MaxValue));
-                _outputSampleCount++;
-                _nextOutputSourcePosition = _outputSampleCount * (double)_sourceRate / _outputRate;
-            }
-
-            _sourceFramesWritten = sliceEnd;
-        }
-
-        public void Complete()
-        {
-            if (_outputSampleCount == 0)
-                throw new SubtitleInputException(Strings.AiSubtitle_NoAudioInRange);
-
-            int dataLength = checked(_outputSampleCount * sizeof(short));
-            using var writer = new BinaryWriter(_stream, Encoding.ASCII, leaveOpen: true);
-            _stream.Position = 4;
-            writer.Write(checked(WaveHeaderLength - 8 + dataLength));
-            _stream.Position = 40;
-            writer.Write(dataLength);
-            _stream.Position = WaveHeaderLength + dataLength;
-        }
-
-        private void WriteHeader()
-        {
-            using var writer = new BinaryWriter(_stream, Encoding.ASCII, leaveOpen: true);
-            writer.Write(Encoding.ASCII.GetBytes("RIFF"));
-            writer.Write(0);
-            writer.Write(Encoding.ASCII.GetBytes("WAVEfmt "));
-            writer.Write(16);
-            writer.Write((short)1);
-            writer.Write((short)1);
-            writer.Write(_outputRate);
-            writer.Write(_outputRate * sizeof(short));
-            writer.Write((short)sizeof(short));
-            writer.Write((short)16);
-            writer.Write(Encoding.ASCII.GetBytes("data"));
-            writer.Write(0);
-        }
-    }
-
-    internal static async Task WriteCaptionExportAsync(
-        IStorageFile file,
-        byte[] bytes,
-        CancellationToken cancellationToken)
-    {
-        Uri path = file.Path;
-        if (path.IsFile && !string.IsNullOrWhiteSpace(path.LocalPath))
-        {
-            string destinationPath = path.LocalPath;
-            string temporaryPath = destinationPath + $".{Guid.NewGuid():N}.tmp";
-            try
-            {
-                UnixFileMode destinationMode = default;
-                var options = new FileStreamOptions
-                {
-                    Mode = FileMode.CreateNew,
-                    Access = FileAccess.Write,
-                    Share = FileShare.None,
-                    Options = FileOptions.WriteThrough,
-                };
-                if (!OperatingSystem.IsWindows())
-                {
-                    destinationMode = File.Exists(destinationPath)
-                        ? File.GetUnixFileMode(destinationPath)
-                        : UnixFileMode.UserRead | UnixFileMode.UserWrite;
-                    options.UnixCreateMode = destinationMode;
-                }
-
-                await using (var stream = new FileStream(temporaryPath, options))
-                {
-                    await stream.WriteAsync(bytes, cancellationToken);
-                    await stream.FlushAsync(cancellationToken);
-                    stream.Flush(flushToDisk: true);
-                }
-
-                if (!OperatingSystem.IsWindows())
-                    File.SetUnixFileMode(temporaryPath, destinationMode);
-                File.Move(temporaryPath, destinationPath, overwrite: true);
-            }
-            finally
-            {
-                try { File.Delete(temporaryPath); }
-                catch (IOException) { }
-                catch (UnauthorizedAccessException) { }
-            }
-        }
-        else
-        {
-            await WriteNonLocalCaptionExportAsync(file, bytes, cancellationToken);
-        }
-    }
-
-    private static async Task WriteNonLocalCaptionExportAsync(
-        IStorageFile destination,
-        byte[] bytes,
-        CancellationToken cancellationToken)
-    {
-        cancellationToken.ThrowIfCancellationRequested();
-        IStorageFolder parent = await destination.GetParentAsync()
-            ?? throw new NotSupportedException(
-                "The storage provider cannot stage a safe replacement for this caption file.");
-        IStorageFolder? stagingFolder = null;
-        IStorageFolder? backupFolder = null;
-        IStorageFile? stagedFile = null;
-        IStorageFile? backupFile = null;
-        bool originalIsInBackup = false;
-        bool preserveRecoveryArtifacts = false;
-        try
-        {
-            string transaction = Guid.NewGuid().ToString("N");
-            stagingFolder = await parent.CreateFolderAsync($"beutl-caption-{transaction}-stage")
-                ?? throw new NotSupportedException(
-                    "The storage provider cannot create a caption staging folder.");
-            stagedFile = await stagingFolder.CreateFileAsync(destination.Name)
-                ?? throw new NotSupportedException(
-                    "The storage provider cannot create a staged caption file.");
-
-            await using (Stream stream = await stagedFile.OpenWriteAsync())
-            {
-                await stream.WriteAsync(bytes, cancellationToken);
-                await stream.FlushAsync(cancellationToken);
-            }
-            cancellationToken.ThrowIfCancellationRequested();
-
-            backupFolder = await parent.CreateFolderAsync($"beutl-caption-{transaction}-backup")
-                ?? throw new NotSupportedException(
-                    "The storage provider cannot create a caption backup folder.");
-            IStorageItem? movedOriginal;
-            try
-            {
-                movedOriginal = await destination.MoveAsync(backupFolder);
-            }
-            catch
-            {
-                preserveRecoveryArtifacts = true;
-                throw;
-            }
-            if (movedOriginal is not IStorageFile movedOriginalFile)
-            {
-                movedOriginal?.Dispose();
-                preserveRecoveryArtifacts = true;
-                throw new IOException("The storage provider could not move the previous caption file to safety.");
-            }
-            backupFile = movedOriginalFile;
-            originalIsInBackup = true;
-
-            try
-            {
-                IStorageItem? movedStaged = await stagedFile.MoveAsync(parent);
-                if (movedStaged is not IStorageFile committed)
-                {
-                    movedStaged?.Dispose();
-                    throw new IOException("The storage provider could not publish the staged caption file.");
-                }
-                TryDisposeStorageItem(committed);
-                originalIsInBackup = false;
-            }
-            catch (Exception commitException)
-            {
-                Exception? rollbackException = await TryRestoreNonLocalCaptionAsync(
-                    parent,
-                    backupFile,
-                    destination.Name);
-                if (rollbackException is not null)
-                {
-                    preserveRecoveryArtifacts = true;
-                    throw new AggregateException(
-                        "Caption export failed and the previous file could not be restored; its backup was preserved.",
-                        commitException,
-                        rollbackException);
-                }
-
-                originalIsInBackup = false;
-                throw;
-            }
-
-            await TryDeleteStorageItemAsync(backupFile);
-        }
-        finally
-        {
-            if (!preserveRecoveryArtifacts)
-            {
-                await TryDeleteStorageItemAsync(stagedFile);
-                if (!originalIsInBackup)
-                {
-                    await TryDeleteStorageItemAsync(stagingFolder);
-                    await TryDeleteStorageItemAsync(backupFolder);
-                }
-            }
-
-            TryDisposeStorageItem(backupFile);
-            TryDisposeStorageItem(stagedFile);
-            TryDisposeStorageItem(backupFolder);
-            TryDisposeStorageItem(stagingFolder);
-            TryDisposeStorageItem(parent);
-        }
-    }
-
-    private static async Task<Exception?> TryRestoreNonLocalCaptionAsync(
-        IStorageFolder parent,
-        IStorageFile backupFile,
-        string destinationName)
-    {
-        try
-        {
-            IStorageFile? failedReplacement = await parent.GetFileAsync(destinationName);
-            try
-            {
-                if (failedReplacement is not null)
-                {
-                    await failedReplacement.DeleteAsync();
-                }
-            }
-            finally
-            {
-                TryDisposeStorageItem(failedReplacement);
-            }
-
-            IStorageItem restored = await backupFile.MoveAsync(parent)
-                ?? throw new IOException("The storage provider returned no restored caption file.");
-            TryDisposeStorageItem(restored);
-            return null;
-        }
-        catch (Exception ex)
-        {
-            return ex;
-        }
-    }
-
-    private static async Task TryDeleteStorageItemAsync(IStorageItem? item)
-    {
-        if (item is null)
-            return;
-
-        try
-        {
-            await item.DeleteAsync();
-        }
-        catch
-        {
-            // A completed replacement must not be reported as failed only because a provider could
-            // not remove its empty staging folder. If rollback failed, recovery artifacts are kept.
-        }
-    }
-
-    private static void TryDisposeStorageItem(IDisposable? item)
-    {
-        try
-        {
-            item?.Dispose();
-        }
-        catch
-        {
-            // Storage handles are cleanup-only here; the write or rollback outcome is authoritative.
-        }
-    }
-
     private FilePickerFileType CreateCaptionFileType(
         bool canDecode = false,
         bool canEncode = false)
@@ -4632,18 +3927,6 @@ public sealed partial class AiSubtitleDialogViewModel
         }
         return TopLevel.GetTopLevel(window)?.StorageProvider;
     }
-
-    private sealed record TranslationPiece(
-        int CueIndex,
-        int PartIndex,
-        int ContextPartIndex,
-        string Id,
-        string GroupId,
-        TimeSpan Start,
-        TimeSpan End,
-        string Text);
-
-    private sealed record TranslationBatch(IReadOnlyList<TranslationPiece> Pieces);
 
     private sealed class TranslationOperation(
         CaptionDocument sourceDocument,
@@ -4849,28 +4132,6 @@ public sealed partial class AiSubtitleDialogViewModel
     }
 }
 
-internal readonly record struct SpeechWaveChunkResult(
-    int SourceSampleCount,
-    int OutputSampleCount,
-    TimeSpan UploadedDuration);
-
 internal readonly record struct SourceAudioFingerprint(
     long FileLength,
     long LastWriteTimeUtcTicks);
-
-internal sealed class SubtitleInputException : Exception
-{
-    public SubtitleInputException()
-    {
-    }
-
-    public SubtitleInputException(string message)
-        : base(message)
-    {
-    }
-
-    public SubtitleInputException(string message, Exception innerException)
-        : base(message, innerException)
-    {
-    }
-}
